pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-triage-env"
version = "1.0.0"
description = "OpenEnv environment for email triage — classify, prioritise, route, and respond to emails"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Team Ctrl-Alt-Defeat"},
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
openenv = [
    "openenv-core>=0.1.0",
]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
]

[tool.setuptools.packages.find]
include = ["server*", "models*"]
